+++ .travis.yml
@@ -8,9 +8,9 @@ script:
 
 deploy:
   provider: pages
-  skip_cleanup: true
+  cleanup: false
   github_token: $GITHUB_TOKEN  # Set in the settings page of your repository, as a secure variable
   keep_history: true
   local_dir: dist
   on:
-    all_branch: true
+    all_branches: true
