+++ example/slipIntroduction.html
@@ -4,8 +4,8 @@
 	<title>Slip slip presentation</title>
 	<meta charset="utf-8" />
 	<link rel="stylesheet" type="text/css" href="slipIntroduction.css">
-	<link rel="stylesheet" type="text/css" href="../slip.css">
-	<link rel="stylesheet" type="text/css" href="../theorem.css">
+	<link rel="stylesheet" type="text/css" href="../src/css/slip.css">
+	<link rel="stylesheet" type="text/css" href="../src/css/theorem.css">
 	<!-- For the use of mathjax -->
 	<!-- <script type="text/javascript" async
 	     src="https://cdnjs.cloudflare.com/ajax/libs/mathjax/2.7.5/MathJax.js?config=TeX-MML-AM_CHTML">
@@ -320,7 +320,7 @@ presentation.setSlips([intro, devel, intro]);</code></pre>
 		<div style="margin: 100px">An example (old version of the engine): <a href="https://choum.net/panglesd/slides/slides-js/slides.html">Slips</a>.</div>
 </div>
 	</div>
-	<script src="slipIntroduction.js"></script>
+	<script src="./slipIntroduction.js"></script>
 	<script>hljs.initHighlightingOnLoad();</script>
     </body>
 </html>
