+++ package.json
@@ -8,7 +8,7 @@
   "license": "MIT",
   "scripts": {
     "dev": "parcel example/slipIntroduction.html",
-    "build": "parcel build example/slipIntroduction.html lib/slip-lib.js"
+    "build": "parcel build example/slipIntroduction.html src/slip-lib.js --public-url /slip-js/"
   },
   "devDependencies": {
     "parcel-bundler": "^1.12.4"
