+++ src/modules/slip.mjs
