+++ src/slip-lib.js
